+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES for `@es-joy/jsdoc-eslint-parser`
 
+## 0.11.1
+
+- fix: avoid adding `jsdocBlocks` to any node except `Program` and drop
+    `jsdoc` from `Program`
+
 ## 0.11.0
 
 - BREAKING CHANGE: chore: update `jsdoccomment`, `typescript-eslint/parser`;